EES-1127 Cleanup download page components
This change:

- Renames `PublicationDownloadList` to `TopicDownloadList` and inlines
  the topic's `Details` into it. This simplifies the loops needed in
  `DownloadIndexPage`.
- Removes casing on file types/sizes on download links to be consistent
  with the release page's download links.
- Removes redundant 'Download files for' text from each publication
  heading.
- Removes unnecessary spacing between download links.
- Removes any unnecessary props on elements e.g. redundant class names.
