+++ src/explore-education-statistics-frontend/src/modules/download/DownloadIndexPage.tsx
@@ -1,15 +1,14 @@
+import Accordion from '@common/components/Accordion';
 import AccordionSection from '@common/components/AccordionSection';
-import Details from '@common/components/Details';
 import RelatedInformation from '@common/components/RelatedInformation';
 import themeService, { DownloadTheme } from '@common/services/themeService';
 import Link from '@frontend/components/Link';
 import Page from '@frontend/components/Page';
 import PageSearchFormWithAnalytics from '@frontend/components/PageSearchFormWithAnalytics';
+import TopicDownloadList from '@frontend/modules/download/components/TopicDownloadList';
+import { logEvent } from '@frontend/services/googleAnalyticsService';
 import { GetServerSideProps, NextPage } from 'next';
 import React from 'react';
-import Accordion from '@common/components/Accordion';
-import { logEvent } from '@frontend/services/googleAnalyticsService';
-import PublicationDownloadList from './components/PublicationDownloadList';
 
 interface Props {
   themes: DownloadTheme[];
@@ -17,14 +16,17 @@ interface Props {
 
 const DownloadIndexPage: NextPage<Props> = ({ themes = [] }) => {
   return (
-    <Page title="Download latest data files" breadcrumbLabel="Download">
+    <Page
+      title="Download latest data files"
+      breadcrumbLabel="Download latest data files"
+    >
       <div className="govuk-grid-row">
         <div className="govuk-grid-column-two-thirds">
           <p className="govuk-body-l">
             Find the latest data files behind our range of national and regional
             statistics for your own analysis.
           </p>
-          <p className="govuk-body">
+          <p>
             Previous release data can be found on their respective release
             pages.
           </p>
@@ -61,28 +63,17 @@ const DownloadIndexPage: NextPage<Props> = ({ themes = [] }) => {
             );
           }}
         >
-          {themes.map(
-            ({
-              id: themeId,
-              title: themeTitle,
-              summary: themeSummary,
-              topics,
-            }) => (
-              <AccordionSection
-                key={themeId}
-                heading={themeTitle}
-                caption={themeSummary}
-              >
-                {topics.map(
-                  ({ id: topicId, title: topicTitle, publications }) => (
-                    <Details key={topicId} summary={topicTitle}>
-                      <PublicationDownloadList publications={publications} />
-                    </Details>
-                  ),
-                )}
-              </AccordionSection>
-            ),
-          )}
+          {themes.map(theme => (
+            <AccordionSection
+              key={theme.id}
+              heading={theme.title}
+              caption={theme.summary}
+            >
+              {theme.topics.map(topic => (
+                <TopicDownloadList key={topic.id} topic={topic} />
+              ))}
+            </AccordionSection>
+          ))}
         </Accordion>
       ) : (
         <div className="govuk-inset-text">No data currently published.</div>
